Long_term_memory_chatbots/Bot with temp chat history summary and multi user vectordb/requirements.txt
streamlit
ollama 
langchain-ollama
langchain_core
langchain_community
faiss-cpu
langchain
python-dotenv
tiktoken